Make pebble config accessible to components.
This refactors the Pebble config from being accessible only to the
toplevel cmds and moves it such that it can be passed directly to the
components (e.g. the VA) to reduce the number of parameters being added
to constructors.
