+++ cmd/pebble-client/main.go
@@ -16,6 +16,7 @@ import (
 	"strings"
 
 	"github.com/letsencrypt/pebble/cmd"
+
 	"gopkg.in/square/go-jose.v2"
 )
